changing the name to ZedDoom and removing certain past contributors' crappy code and messy dependency handling

+++ src/sound/music_win_mididevice.cpp
@@ -175,7 +175,7 @@ int WinMIDIDevice::GetTechnology() const
 
 int WinMIDIDevice::SetTempo(int tempo)
 {
-	MIDIPROPTEMPO data = { sizeof(MIDIPROPTEMPO), (DWORD)tempo };
+	MIDIPROPTEMPO data = { sizeof(MIDIPROPTEMPO), tempo };
 	return midiStreamProperty(MidiOut, (LPBYTE)&data, MIDIPROP_SET | MIDIPROP_TEMPO);
 }
 
@@ -187,7 +187,7 @@ int WinMIDIDevice::SetTempo(int tempo)
 
 int WinMIDIDevice::SetTimeDiv(int timediv)
 {
-	MIDIPROPTIMEDIV data = { sizeof(MIDIPROPTIMEDIV), (DWORD)timediv };
+	MIDIPROPTIMEDIV data = { sizeof(MIDIPROPTIMEDIV), timediv };
 	return midiStreamProperty(MidiOut, (LPBYTE)&data, MIDIPROP_SET | MIDIPROP_TIMEDIV);
 }
